Fix bug where read-only deck revision shares same subdeck with latest
When adding a subdeck and then reverting, the subdeck is kept in the
older (now read-only) revision of its parent without the subdeck being
revised.

If at a later time we revert to that reverted revision, if we don't
create a revision for the subdeck we end up having the same subdeck
revision in two parent deck revisions, one of which is read-only, and
the other isn't.

This violates the principle that for any deck revision other than the
last, the deck tree should remain read-only as a whole, i.e. it
should only include read-only revisions.