Refactor slide/deck translation to use new simple translation API
The previous implementation used a silly API where the deck service
would ask translation from translation service, then the translation
service would ask the deck service for content to translate, and
translated objects were returned to the deck service to include in
the translated deck.

Now we simply send whatever text we want to the translation service
and the deck service handles the rest.

Sort of.